When calling to_string and to_string_binary for double values, they get casted to floats and float versions of functions get called, this is very bad but it is ok for now. I am not expecting values that are to big for float at this time.

+++ src/kernel.cpp
@@ -54,7 +54,7 @@ void kernel_main()
     string str5 = StringBuilder::FormatString(
         "sign: {b} \ndisplacment: {} \nwhole_part: {}\nfractional_part : {} \npow_test: {}",
         fp.sign,
-        fp.displacment, fp.whole_part, 1, (float) 0.0012);
+        fp.displacment, fp.whole_part, 1, 0.0012);
     print(str5.getCharPointer(), str5.size());
 
     // print(str4.getCharPointer(), str4.size());
